% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utility_functions.R
\name{download_data}
\alias{download_data}
\title{Download SDP Datasets Locally}
\usage{
download_data(
  urls,
  output_dir,
  return_status = TRUE,
  resume = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{urls}{character vector of URLs for files to download}

\item{output_dir}{character vector of destination file paths or single path for output files.}

\item{return_status}{logical. Should the function return a data frame with results from the download session?}

\item{resume}{logical. Should the function resume partial downloads?}

\item{overwrite}{logical. Should the function overwrite existing files?}

\item{...}{arguments passed to \code{curl::multi_download()}}
}
\description{
Download SDP Datasets Locally
}
