% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_catalog_functions.R
\name{get_sdp_catalog}
\alias{get_sdp_catalog}
\title{Discover SDP data by downloading catalog information.}
\usage{
get_sdp_catalog(
  domains = c("UG", "UER", "GT"),
  types = c("Mask", "Topo", "Vegetation", "Planning", "Radiation", "Snow", "Imagery",
    "Supplemental"),
  releases = c("Basemaps", "Release1", "Release2", "Release3"),
  deprecated = FALSE,
  return_stac = FALSE
)
}
\arguments{
\item{domains}{The spatial domain of the desired data product.}

\item{types}{The type of product to return.}

\item{releases}{Which release (group) of data products to return.}

\item{deprecated}{Should older versions of datasets be returned, or just the latest version?}

\item{return_stac}{NOT IMPLEMENTED: Should the results be returned as a Spatio-temporal Asset Catalog? Otherwise return and ordinary data frame.}
}
\value{

}
\description{
Discover SDP data by downloading catalog information.
}
\examples{

## Gets information on all current datasets.
sdp_cat <- get_sdp_catalog()
str(sdp_cat)

## Gets a subset of catalog entries for the Upper Gunnison (UG) domain.
sdp_sub <- get_sdp_catalog(domains="UG", types="Vegetation",
                          deprecated=FALSE,return_stac=FALSE)
sdp_sub

}
