% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_catalog_functions.R
\name{get_sdp_metadata}
\alias{get_sdp_metadata}
\title{Download geospatial metadata for SDP Datasets.}
\usage{
get_sdp_metadata(catalog_id, return_list = TRUE)
}
\arguments{
\item{catalog_id}{The unique Catalog ID code for the desired dataset.}

\item{return_list}{Logical. If \code{TRUE}, then the output is parsed as an R list object.}
}
\value{
A nested list (\code{return_list=TRUE}) or XML document (\code{return_list=FALSE}) containing detailed geospatial metadata for each dataset.
}
\description{
Download geospatial metadata for SDP Datasets.
}
\examples{

##Get metadata for a specific item.
get_sdp_metadata(catalog_id="R1D001")


}
