% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_utility_functions.R
\name{replace_strngs}
\alias{replace_strngs}
\title{Replace Multiple Strings in a Vector}
\usage{
replace_strngs(x, y, vec, ...)
}
\arguments{
\item{x}{vector with strings to replace}

\item{y}{vector with strings to use instead}

\item{vec}{initial character vector}

\item{...}{arguments passed to \code{gsub}}
}
\description{
Replace Multiple Strings in a Vector
}
