% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSDP-package.R
\docType{package}
\name{rSDP-package}
\alias{rSDP}
\alias{rSDP-package}
\title{rSDP: Discover, Query, and Subset Data from the RMBL Spatial Data Platform}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The rSDP package provides a simplified interface for discovering, querying, and subsetting data products that are incorporated into the RMBL Spatial Data Platform. The RMBL SDP provides a set of curated, high-resolution, and high-fidelity geospatial datasets for a set of domains in Western Colorado (USA) in the vicinity of [Rocky Mountain Biological Laboratory](https://rmbl.org). For more information about the RMBL SDP [see here](https://www.rmbl.org/scientists/resources/spatial-data-platform/). SDP data products are provided as geospatial raster datasets in [cloud-optimized Geotiff]() (COG) format. The rSDP package provides functions to access these datasets in cloud storage (Amazon S3) without downloading.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmbl-sdp/rSDP}
  \item \url{https://rmbl-sdp.github.io/rSDP}
  \item Report bugs at \url{https://github.com/rmbl-sdp/rSDP/issues}
}

}
\author{
\strong{Maintainer}: Ian Breckheimer \email{ikb@rmbl.org} (\href{https://orcid.org/0000-0002-4698-977X}{ORCID})

}
\keyword{internal}
