% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_catalog_functions.R
\name{sdp_get_catalog}
\alias{sdp_get_catalog}
\title{Discover SDP data by downloading catalog information.}
\usage{
sdp_get_catalog(
  domains = .SDP_DOMAINS,
  types = .SDP_TYPES,
  releases = .SDP_RELEASES,
  timeseries_types = .SDP_TIMESERIES_TYPES,
  deprecated = FALSE,
  return_stac = FALSE
)
}
\arguments{
\item{domains}{The spatial domain of the desired data product.}

\item{types}{The type of product to return.}

\item{releases}{Which release (group) of data products to return.}

\item{timeseries_types}{Some datasets are structured as single datasets (e.g. \code{Single}), and others are time-series with various periods (e.g.\code{Monthly}).}

\item{deprecated}{Should older versions of datasets be returned, or just the latest version?}

\item{return_stac}{Logical. If \code{TRUE}, returns the root of the RMBL SDP's static STAC catalog as an \code{rstac} object (requires the \code{rstac} package). The returned catalog is browseable via \code{rstac} traversal methods but does not support \code{rstac::stac_search()} (which requires a STAC API server). You can browse the catalog interactively at \verb{https://radiantearth.github.io/stac-browser/#/external/rmbl-sdp.s3.us-east-2.amazonaws.com/stac/v1/catalog.json}. When \code{return_stac=TRUE}, filter arguments (\code{domains}, \code{types}, etc.) are ignored; use \code{rstac} methods to filter the returned catalog. Default \code{FALSE} returns an ordinary data frame.}
}
\value{
A data frame containing basic catalog information for the matched data products.
}
\description{
Discover SDP data by downloading catalog information.
}
\examples{

## Gets information on all current datasets.
sdp_cat <- sdp_get_catalog()
str(sdp_cat)

## Gets a subset of catalog entries for the Upper Gunnison (UG) domain.
sdp_sub <- sdp_get_catalog(domains="UG", types="Vegetation",
                          deprecated=FALSE,return_stac=FALSE)
sdp_sub

}
