% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_catalog_functions.R
\name{sdp_get_catalog}
\alias{sdp_get_catalog}
\title{Discover SDP data by downloading catalog information.}
\usage{
sdp_get_catalog(
  domains = c("UG", "UER", "GT"),
  types = c("Mask", "Topo", "Vegetation", "Hydro", "Planning", "Radiation", "Snow",
    "Imagery", "Supplemental"),
  releases = c("Basemaps", "Release1", "Release2", "Release3", "Release4"),
  timeseries_types = c("Single", "Annual", "Seasonal", "Monthly", "Daily"),
  deprecated = FALSE,
  return_stac = FALSE
)
}
\arguments{
\item{domains}{The spatial domain of the desired data product.}

\item{types}{The type of product to return.}

\item{releases}{Which release (group) of data products to return.}

\item{timeseries_types}{Some datasets are structured as single datasets (e.g. \code{Single}), and others are time-series with various periods (e.g.\code{Monthly}).}

\item{deprecated}{Should older versions of datasets be returned, or just the latest version?}

\item{return_stac}{NOT IMPLEMENTED: Should the results be returned as a Spatio-temporal Asset Catalog? Otherwise return and ordinary data frame.}
}
\value{
A data frame containing basic catalog information for the matched data products.
}
\description{
Discover SDP data by downloading catalog information.
}
\examples{

## Gets information on all current datasets.
sdp_cat <- sdp_get_catalog()
str(sdp_cat)

## Gets a subset of catalog entries for the Upper Gunnison (UG) domain.
sdp_sub <- sdp_get_catalog(domains="UG", types="Vegetation",
                          deprecated=FALSE,return_stac=FALSE)
sdp_sub

}
