% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_extraction_functions.R
\name{sdp_get_raster}
\alias{sdp_get_raster}
\title{Create an R object representing an SDP cloud-based dataset.}
\usage{
sdp_get_raster(
  catalog_id = NULL,
  url = NULL,
  years = NULL,
  date_start = NULL,
  date_end = NULL,
  ...
)
}
\arguments{
\item{catalog_id}{character. A single valid catalog number for an SDP dataset. This is in the \code{CatalogID} field for information returned by \code{sdp_get_catalog()}.}

\item{url}{character. A valid URL (e.g. https://path.to.dataset.tif) for the cloud-based dataset. You should specify either \code{catalog_id} or \code{url}, but not both.}

\item{years}{numeric. For annual time-series data, a numeric vector specifying which years to return. The default \code{NULL} returns all available years.}

\item{date_start}{class \code{Date}. For daily time-series data, the first day of data to return.}

\item{date_end}{class \code{Date}. For daily time-series data, the last day of data to return.}

\item{...}{Other arguments to pass to the \code{terra::rast()} function.}
}
\value{
An R object (class \code{terra::SpatRaster}) representing the raster dataset.
}
\description{
Create an R object representing an SDP cloud-based dataset.
}
\details{
Files headers are read from cloud-based datasets using the \code{terra} package, but the full dataset is not downloaded locally. Instead \code{terra} uses the web-based file system embedded in GDAL (VSICURL) to access datasets on the cloud. For large datasets and slow network connections, the function might take up to a minute to complete.
}
\examples{
## Lookup catalog number for a dataset.
cat <- sdp_get_catalog(domain='UG',type='Vegetation')
lc_id <- cat$CatalogID[cat$Product=='Basic Landcover']

## Connect to the dataset without downloading
landcover <- sdp_get_raster(lc_id)
landcover

}
